// src/lib/surf/types.ts
export type SurfForecastData = {
  swimRisk: string | undefined;
  waveHeight: string | undefined;
  wavePeriod: string | undefined;

  uvIndex: string | undefined;
  waterTemp: string | undefined;

  weather: string | undefined;

  highTemp: string | undefined;
  winds: string | undefined;

  sunrise: string | undefined;
  sunset: string | undefined;
};